refactor: moving common functionality to Connector
